SGXP-863 move BadgeDrawable to icon package
