+++ designsystem/src/main/kotlin/com/natura/android/appbar/SetupAppBar.kt
@@ -2,7 +2,7 @@ package com.natura.android.appbar
 
 import android.app.Activity
 import android.view.Menu
-import com.natura.android.R
+import com.natura.android.icon.BadgeDrawable
 
 class SetupAppBar{
 
@@ -20,18 +20,12 @@ class SetupAppBar{
             configureOptionsMenu(activity, menu, appbarMenu)
 
             menu.findItem(iconToFind)?.let { menuItem ->
-                badgeDrawable = BadgeDrawable(activity, mCount, menuItem.icon )
+                badgeDrawable = BadgeDrawable(
+                    activity,
+                    mCount,
+                    menuItem.icon
+                )
             }}
-    }
-
-     fun displayMenu(
-        activity: Activity,
-        menu: Menu?,
-        appbarMenu: Int
-    ) {
-        menu?.let {
-            configureOptionsMenu(activity, it, appbarMenu)
-            it.findItem(R.id.ic_schedule).let { menuItem ->  menuItem.setIcon(R.drawable.ds_button_primary_enabled)}}
     }
 
     private fun configureOptionsMenu(activity: Activity, menu: Menu, appbarMenu: Int) {
